requirements.txt
Flask
requests
beautifulsoup4
celery
redis
supabase
flask-cors
werkzeug
python-dotenv
gunicorn